Refactor settings pages
Re-implemented the base page using react-router capabilities.
Moved the settings navigation to the right column.
Moved ProfileHeader from the base page into the SettingsPage.

I had to ditch the implementation of the settings forms using
Command (utils/command.js) because it did't work with the new
base page. There are remains of the command system here and
there that should be removed/utilized in the future.